# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RetailIQ API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/profile": {
      "get": {
        "summary": "get_profile",
        "operationId": "get_profile_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "update_profile",
        "operationId": "update_profile_put",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "store_type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/{category_id}": {
      "put": {
        "summary": "update_category",
        "operationId": "update_category_put",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "delete_category",
        "operationId": "delete_category_delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/tax-config": {
      "get": {
        "summary": "get_tax_config",
        "operationId": "get_tax_config_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "update_tax_config",
        "operationId": "update_tax_config_put",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taxes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tax/config": {
      "get": {
        "summary": "get_tax_config",
        "operationId": "get_tax_config_get",
        "parameters": [
          {
            "name": "country_code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/tax/calculate": {
      "post": {
        "summary": "calculate_tax",
        "operationId": "calculate_tax_post",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Response 400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Response 500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "string"
                  },
                  "country_code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tax/filing-summary": {
      "get": {
        "summary": "tax_summary",
        "operationId": "tax_summary_get",
        "parameters": [
          {
            "name": "country_code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}